Widen proto-local symbol indices

+++ src/module/xmodule.c
@@ -316,13 +316,13 @@ void xr_module_build_export_index(XrModule *module) {
 
     module->min_symbol = min_sym;
     module->max_symbol = max_sym;
-    module->symbol_to_index = (int16_t *) xr_malloc(range * sizeof(int16_t));
+    module->symbol_to_index = (int32_t *) xr_malloc(range * sizeof(int32_t));
     if (!module->symbol_to_index)
         return;
-    memset(module->symbol_to_index, 0xFF, range * sizeof(int16_t));  // -1 (0xFFFF)
+    memset(module->symbol_to_index, 0xFF, range * sizeof(int32_t));  // -1
 
     for (uint16_t i = 0; i < module->export_count; i++) {
-        module->symbol_to_index[module->export_symbols[i] - min_sym] = (int16_t) i;
+        module->symbol_to_index[module->export_symbols[i] - min_sym] = (int32_t) i;
     }
 }
